test: extract value commit tests into a separate file
