+++ packages/date-time-picker/test/basic.test.js
@@ -247,360 +247,6 @@ describe('Basic features', () => {
     });
   });
 
-  describe('change event', () => {
-    const changeEventType = 'change';
-    const unparsableChangeEventType = 'unparsable-change';
-    let changeEventSpy;
-    let unparsableChangeEventSpy;
-
-    beforeEach(() => {
-      changeEventSpy = sinon.spy();
-      dateTimePicker.addEventListener(changeEventType, changeEventSpy);
-      unparsableChangeEventSpy = sinon.spy();
-      dateTimePicker.addEventListener(unparsableChangeEventType, unparsableChangeEventSpy);
-    });
-
-    ['date-picker', 'time-picker'].forEach((pickerType) => {
-      function getPicker(pickerType) {
-        return pickerType === 'date-picker' ? datePicker : timePicker;
-      }
-
-      function getOtherPickerType(pickerType) {
-        return pickerType === 'date-picker' ? 'time-picker' : 'date-picker';
-      }
-
-      function getPickerInitialValue(pickerType) {
-        return pickerType === 'date-picker' ? '2/2/2022' : '02:02';
-      }
-
-      function getPickerNewValue(pickerType) {
-        return pickerType === 'date-picker' ? '1/1/2023' : '13:00';
-      }
-
-      async function triggerInputChange(trigger) {
-        if (trigger === 'outside click') {
-          outsideClick();
-        } else {
-          await sendKeys({ press: 'Enter' });
-        }
-      }
-
-      async function clearValueOnFocusedInput(trigger) {
-        await sendKeys({ press: 'ControlOrMeta+A' });
-        await sendKeys({ press: 'Backspace' });
-        await triggerInputChange(trigger);
-        await nextRender();
-      }
-
-      async function changeValueOnFocusedInput(newValue, trigger) {
-        await sendKeys({ press: 'ControlOrMeta+A' });
-        await sendKeys({ type: newValue });
-        await triggerInputChange(trigger);
-        await nextRender();
-      }
-
-      async function initializePickerState(pickerType, initialState) {
-        if (initialState === 'empty') {
-          return;
-        }
-        const picker = getPicker(pickerType);
-        picker.focus();
-        const value = initialState === 'parsable' ? getPickerInitialValue(pickerType) : 'unparsableInitial';
-        await changeValueOnFocusedInput(value);
-        picker.blur();
-      }
-
-      async function changeStateOnFocusedInput(pickerType, newPickerState, trigger) {
-        if (newPickerState === 'empty') {
-          await clearValueOnFocusedInput(trigger);
-        } else {
-          const newValue = newPickerState === 'unparsable' ? 'unparsableNew' : getPickerNewValue(pickerType);
-          await changeValueOnFocusedInput(newValue, trigger);
-        }
-      }
-
-      async function initializeAndFocus(pickerType, pickerInitialState, otherPickerInitialState) {
-        await initializePickerState(pickerType, pickerInitialState);
-        await initializePickerState(getOtherPickerType(pickerType), otherPickerInitialState);
-        changeEventSpy.resetHistory();
-        unparsableChangeEventSpy.resetHistory();
-        getPicker(pickerType).focus();
-      }
-
-      it(`should fire change on ${pickerType} outside click when value is changed from empty to parsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'outside click');
-        expect(changeEventSpy.called).to.be.true;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire change on ${pickerType} outside click when value is changed from parsable to empty when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'outside click');
-        expect(changeEventSpy.called).to.be.true;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from empty to unparsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from unparsable to empty when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire change on ${pickerType} outside click when value is changed from parsable to unparsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'outside click');
-        expect(changeEventSpy.called).to.be.true;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire change on ${pickerType} outside click when value is changed from unparsable to parsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'outside click');
-        expect(changeEventSpy.called).to.be.true;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from empty to parsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from parsable to empty when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from empty to unparsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from unparsable to empty when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from parsable to unparsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from unparsable to parsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from empty to parsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from parsable to empty when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from empty to unparsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from unparsable to empty when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from parsable to unparsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} outside click when value is changed from unparsable to parsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'outside click');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire change on ${pickerType} enter when value is changed from empty to parsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'enter');
-        expect(changeEventSpy.called).to.be.true;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should not fire any change events on ${pickerType} enter when value is changed from parsable to empty when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from empty to unparsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from unparsable to empty when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire change on ${pickerType} enter when value is changed from parsable to unparsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'enter');
-        expect(changeEventSpy.called).to.be.true;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire change on ${pickerType} enter when value is changed from unparsable to parsable when ${getOtherPickerType(pickerType)} value is parsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'parsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'enter');
-        expect(changeEventSpy.called).to.be.true;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from empty to parsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from parsable to empty when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from empty to unparsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from unparsable to empty when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from parsable to unparsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from unparsable to parsable when ${getOtherPickerType(pickerType)} value is unparsable`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'unparsable');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should not fire any change events on ${pickerType} enter when value is changed from empty to parsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.false;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from parsable to empty when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from empty to unparsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'empty', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from unparsable to empty when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'empty', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from parsable to unparsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'parsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'unparsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-
-      it(`should fire unparsable-change on ${pickerType} enter when value is changed from unparsable to parsable when ${getOtherPickerType(pickerType)} value is empty`, async () => {
-        await initializeAndFocus(pickerType, 'unparsable', 'empty');
-        await changeStateOnFocusedInput(pickerType, 'parsable', 'enter');
-        expect(changeEventSpy.called).to.be.false;
-        expect(unparsableChangeEventSpy.called).to.true;
-      });
-    });
-
-    it('should not fire change on programmatic value change', () => {
-      dateTimePicker.value = '2020-01-17T16:00';
-      expect(changeEventSpy.called).to.be.false;
-    });
-
-    it('should not fire change on programmatic value change after manual one', () => {
-      dateTimePicker.value = '2020-01-17T16:00'; // Init with valid value
-      changeInputValue(datePicker, '2020-01-20');
-      changeEventSpy.resetHistory();
-      dateTimePicker.value = '2020-01-10T12:00';
-      expect(changeEventSpy.called).to.be.false;
-    });
-
-    it('should not fire change on programmatic value change after partial manual one', () => {
-      changeInputValue(datePicker, '2020-01-17');
-      // Time picker has no value so date time picker value is still empty
-      dateTimePicker.value = '2020-01-17T16:00';
-      expect(changeEventSpy.called).to.be.false;
-    });
-  });
-
   describe('value property formats', () => {
     it('should accept ISO format', () => {
       const date = new Date(Date.UTC(0, 1, 3, 8, 30, 0));
